requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0
streamlit>=1.30.0
plotly>=5.18.0
gspread>=5.12.0
google-auth>=2.23.0